Use rest client
