+++ core/src/main/java/hudson/PluginManager.java
@@ -95,6 +95,7 @@ import org.kohsuke.stapler.StaplerOverridable;
 import org.kohsuke.stapler.StaplerProxy;
 import org.kohsuke.stapler.StaplerRequest;
 import org.kohsuke.stapler.StaplerResponse;
+import org.kohsuke.stapler.bind.JavaScriptMethod;
 import org.kohsuke.stapler.export.Exported;
 import org.kohsuke.stapler.export.ExportedBean;
 import org.kohsuke.stapler.interceptor.RequirePOST;
@@ -155,6 +156,8 @@ import java.util.stream.Collectors;
 
 import static hudson.init.InitMilestone.*;
 import static java.util.logging.Level.*;
+import static java.util.stream.Collectors.toList;
+
 import org.springframework.security.core.Authentication;
 
 /**
@@ -1344,6 +1347,105 @@ public abstract class PluginManager extends AbstractModelObject implements OnMas
         LogFactory.release(uberClassLoader);
     }
 
+    @Restricted(NoExternalUse.class)
+    public static boolean isNonMetaLabel(String label) {
+        return !("adopt-this-plugin".equals(label) || "deprecated".equals(label));
+    }
+    
+    @Restricted(NoExternalUse.class)
+    public HttpResponse doPluginsSearch(@QueryParameter String query, @QueryParameter Integer limit) {
+        String lowerSearchQuery = query != null ? query.toLowerCase() : null;
+        List<JSONObject> plugins = new ArrayList<>();
+        for (UpdateSite site : Jenkins.get().getUpdateCenter().getSiteList()) {
+            plugins = site.getAvailables().stream()
+                .filter(plugin -> {
+                    if (StringUtils.isBlank(query)) {
+                        return true;
+                    }
+                    return plugin.name.toLowerCase().contains(lowerSearchQuery) ||
+                        plugin.title.toLowerCase().contains(lowerSearchQuery) ||
+                        plugin.excerpt.toLowerCase().contains(lowerSearchQuery) ||
+                        Arrays.asList(plugin.categories).contains(query) ||
+                        Arrays.stream(plugin.categories)
+                            .map(UpdateCenter::getCategoryDisplayName)
+                            .anyMatch(category -> category.toLowerCase().contains(query)) ||
+                        plugin.hasWarnings() && lowerSearchQuery.equals("warning:");
+                })
+                .limit(limit)
+                .map(plugin -> {
+                    JSONObject jsonObject = new JSONObject();
+                    jsonObject.put("name", plugin.name);
+                    jsonObject.put("sourceId", plugin.sourceId);
+                    jsonObject.put("title", plugin.title);
+                    jsonObject.put("displayName", plugin.getDisplayName());
+                    jsonObject.put("wiki", plugin.wiki);
+                    jsonObject.put("categories", Arrays.stream(plugin.categories)
+                        .filter(PluginManager::isNonMetaLabel)
+                        .map(UpdateCenter::getCategoryDisplayName)
+                        .collect(toList())
+                    );
+
+                    if (hasAdoptThisPluginLabel(plugin)) {
+                        jsonObject.put("adoptMe", hudson.model.Messages.UpdateCenter_adoptThisPlugin());
+                    }
+                    if (plugin.isDeprecated()) {
+                        jsonObject.put("deprecated", hudson.model.Messages.UpdateCenter_deprecationWarning(plugin.getDeprecation().url));
+                    }
+                    jsonObject.put("excerpt", plugin.excerpt);
+                    jsonObject.put("version", plugin.version);
+                    if (plugin.isForNewerHudson()) {
+                        jsonObject.put("newerCoreRequired", hudson.model.Messages.UpdateCenter_coreWarning(plugin.requiredCore));
+                    }
+                    if (plugin.isForNewerJava()) {
+                        jsonObject.put("newerJavaRequired", hudson.model.Messages.UpdateCenter_javaWarning(plugin.minimumJavaVersion));
+                    }
+                    if (plugin.isNeededDependenciesForNewerJava()) {
+                        VersionNumber javaVersion = plugin.getNeededDependenciesMinimumJavaVersion();
+                        if (javaVersion == null) {
+                            throw new IllegalStateException("java version cannot be null here");
+                        }
+                        jsonObject.put("dependenciesNewerJava", hudson.model.Messages.UpdateCenter_depJavaWarning(javaVersion.toString()));
+                    }
+                    if (plugin.hasWarnings()) {
+                        JSONObject unresolvedSecurityWarnings = new JSONObject();
+                        unresolvedSecurityWarnings.put("text", hudson.model.Messages.UpdateCenter_securityWarning());
+                        Set<UpdateSite.Warning> pluginWarnings = plugin.getWarnings();
+                        if (pluginWarnings == null) {
+                            throw new IllegalStateException("warnings cannot be null here");
+                        }
+                        List<JSONObject> warnings = pluginWarnings.stream()
+                            .map(warning -> {
+                                JSONObject jsonWarning = new JSONObject();
+                                jsonWarning.put("url", warning.url);
+                                jsonWarning.put("message", warning.message);
+                                return jsonWarning;
+                            }).collect(toList());
+                        unresolvedSecurityWarnings.put("warnings", warnings);
+                        jsonObject.put("unresolvedSecurityWarnings", unresolvedSecurityWarnings);
+                    }
+                    if (plugin.releaseTimestamp != null) {
+                        JSONObject releaseTimestamp = new JSONObject();
+                        releaseTimestamp.put("iso8601", Functions.iso8601DateTime(plugin.releaseTimestamp));
+                        releaseTimestamp.put("displayValue", hudson.model.Messages.UpdateCenter_ago(Functions.getTimeSpanString(plugin.releaseTimestamp)));
+                        jsonObject.put("releaseTimestamp", releaseTimestamp);
+                    }
+                    if (hasLatestVersionNewerThanOffered(plugin)) {
+                        jsonObject.put("newerVersionAvailableNotOffered", hudson.model.Messages.UpdateCenter_newerVersionExists(plugin.latest));
+                    }
+                    return jsonObject;
+                })
+                .collect(toList());
+            if (plugins.size() == limit) {
+                break;
+            }
+        }
+
+        JSONArray mappedPlugins = new JSONArray();
+        mappedPlugins.addAll(plugins);
+        
+        return hudson.util.HttpResponses.okJSON(mappedPlugins);
+    }
+    
     /**
      * Get the list of all plugins - available and installed.
      * @return The list of all plugins - available and installed.
