+++ war/src/main/js/api/pluginManager.js
@@ -154,6 +154,22 @@ pluginManager.availablePlugins = function(handler) {
 	});
 };
 
+pluginManager.availablePluginsSearch = function(query, limit, handler) {
+	jenkins.get('/pluginManager/pluginsSearch?query=' + query + '&limit=' + limit, function(response) {
+		if(response.status !== 'ok') {
+			handler.call({ isError: true, errorMessage: response.message });
+			return;
+		}
+
+		handler.call({ isError: false }, response.data);
+	}, {
+		timeout: pluginManagerErrorTimeoutMillis,
+		error: function(xhr, textStatus, errorThrown) {
+			handler.call({ isError: true, errorMessage: errorThrown });
+		}
+	});
+};
+
 
 /**
  * Accepts 1 or 2 arguments, if argument 2 is not provided all installing plugins will be passed
